requirements.txt
Django==3.2.23
djangorestframework==3.14.0
feedparser==6.0.10
requests==2.31.0
python-dateutil==2.8.2
celery==5.3.4
django-celery-beat==2.5.0